+++ pr_agent/tools/pr_questions.py
@@ -2,7 +2,7 @@ import copy
 
 from jinja2 import Environment, StrictUndefined
 
-from pr_agent.algo.ai_handler import BaseAiHandler, LiteLLMAiHandler
+from pr_agent.algo.litellm_ai_handler import BaseAiHandler, LiteLLMAiHandler
 from pr_agent.algo.pr_processing import get_pr_diff, retry_with_fallback_models
 from pr_agent.algo.token_handler import TokenHandler
 from pr_agent.config_loader import get_settings
