+++ pr_agent/tools/pr_reviewer.py
@@ -6,7 +6,7 @@ import yaml
 from jinja2 import Environment, StrictUndefined
 from yaml import SafeLoader
 
-from pr_agent.algo.ai_handler import BaseAiHandler, LiteLLMAiHandler
+from pr_agent.algo.litellm_ai_handler import BaseAiHandler, LiteLLMAiHandler
 from pr_agent.algo.pr_processing import get_pr_diff, retry_with_fallback_models
 from pr_agent.algo.token_handler import TokenHandler
 from pr_agent.algo.utils import convert_to_markdown, load_yaml, try_fix_yaml
